# [Chain][Performance] Rework orphan pruning