@mimirmim 2020-01-02
Going to give this a whirl now and check the improvements over the next couple of days. Thanks for your work on this.