@CaveSpectre11 2020-01-02
> * `CleanBlockIndexGarbage` runs more times than it should be. Upwards to 8 times between blocks.

This isn't changed from the original implementation.  It's called from ProcessNewBlockHeaders().  While it makes sense to move it to somewhere perhaps more appropriate, to reduce the number of calls we would have to investigate why ProcessNewBlockHeaders() is being called so much (I suspect it's due to more "solutions" coming in for a block...  An area that needs some focus anyway, to make sure we're not submitting worse solutions then we already have submitted (for instance); as well as potentially looking if there's a way to triage solutions coming in and getting rid of them sooner if we know it won't be the accepted block.

Given the time spent in the routine is significantly less now, I suspect the reason you see it more frequently is due to that effect... you have more time to do other things, like accept more block headers, since you're not blocked up in there as much :).

While I believe there is significantly more to do, I chose to snap the line on this PR due to the immediate improvements it provides.  I'm not sure tackling a frequency reduction is scope that would be required for this to move into the mainline.

> * Since it cleans up to block 16, does that force the chain to be set in stone after 16 blocks?

This is merely cleaning out what it believes to be stale tips.  I believe that the ancestry check will keep them around longer if they're progressing, and if one does surface as the longest chain, we'll pull the blocks (we deleted) again.   A little bit about the original logic; it pruned when 1000 'irrelevant' blocks were found; regardless of how old they are.  In other words; when that 1000th is determined, regardless if it's one back, or an uncle to the current tip you have, it's going to throw it away.  So basing the purge on height rather than the number of them is (in my opinion) infinitely better.

I'm not sure the best way to test this, and confirm if 12 blocks is the right number.  I'm considering splitting the two, since there will still be significant improvement if we're not going through the ancestry check on over 1/2 million blocks in the current chain; the ancestry check on 2 vs. 1000 shouldn't be very significant from a percentage of the total in mapBlockIndex.

> * It seems to consider the tip to be an "irrelevant index", then deletes it when a new block is produced.

The original code considered any that aren't in the current chain (or passing the ancestor check) as "irrelevant".  The change now is that we wait 12 blocks before we remove it; it will keep coming up as irrelevant on each check until either it becomes not irrelevant, or it becomes part of the active chain.  Watching the debug message reveals that this is the case... sometimes 0 of 2 are removed, and then later 1 of 1 is removed... implying the other one became part of the active chain.

> However, it seems like the bottleneck is now fixed and the RPC calls are super fast again.

This specifically is why I don't want to make riskier changes in this PR, and leave it as is.  However, I just split out the performance change into another PR; since I suspect that is the bulk of the bottleneck (although having upwards of 1000 stale tips may be the actual root cause of the bottleneck, the amount of time locked in that check definitely would have some effect).

#728 should be tested to see how much of an improvement that adds without the question of "is 12 too small of a chain deviation to trim" holding up the merge.