@CaveSpectre11 2020-01-04
I did a lot of rework on this:
1) Moved it from the headers to when we actually attach a new block; which should reduce the frequency of checks.
2) Greatly reworked the conditions for checking.  It now checks when the mapBlockIndex is "bloated".  It's considered bloated when there's 12 or more extra blocks in the map.
3) Reworked the depth of pruning from 12 to 120.  So if it's more than 120 blocks stale, it will prune it then.
4) Added more conditions to what we'll check.  If it's deep enough that we've checked it before, don't bother since it's likely part of the active chain.
5) We set "deep enough" to be the depth (120) below the deepest one we pruned on the last pass.  So the first pass will take a little while (~.5 seconds right now).  The second pass will check fairly deep as well, but most of those should already be cleaned.  Passes after the second pass should be much faster.