@mimirmim 2020-01-06
Testing it out now. Code looks fine, but, I for some reason feel like this may have unforeseen consequences. Give me a day for running this.