@sinetek 2020-01-09
Maybe we should set an upper limit on the number of checks that are performed so as to remain deterministic (and bounded) wrt/ runtime.

| for (const auto& p : mapBlockIndex) {

set an upper limit to the number of items to consider at any one time.
The idea is to clean out over time, it needn't be done in one shot necessarily.