@CaveSpectre11 2020-01-09
> Maybe we should set an upper limit on the number of checks that are performed so as to remain deterministic (and bounded) wrt/ runtime.

The trouble is finding the stale ones.  Reducing the frequency of the passes is what the PRUNE_COUNT is meant to do:

>     // If mapBlockIndex isn't bloated, don't bother taking the time.
>     if (chainActive.Height() > (mapBlockIndex.size() - PRUNE_COUNT)) {
>         return;
>     }

The trouble is the biggest one is actually on startup; there's somewhere on the order of 54,000 stale tips on our disks.  At some point we will figure out how to prune the on disk data so we can clean out the snapshots, but for now that first pass at startup cleans out a significant number.

After that point, the bulk of the time is processing the loop looking for the stale tips.  This (currently) takes 90-120ms (on my machine, with our current 520k blocks).  This of course will grow over time, but I'm predicting 100ms a year give or take. 

So we're at 90-120ms per block received (and linearly growing), only when we're bloated enough, is exponentially better than 1.2-1.4 seconds per header received.

I did take some time to look at where in the map the stale indexes were (after the first cleanup).  I know some places in the code sort the map, so if it was sorted, there was the opportunity to run it until we hit the point where we've already cleared out; knowing the rest of the blocks are in our chain.  However their locations are random; and sorting would not be efficient.

In a nutshell; it's now to a point where the bulk of the time is spent finding the stale blocks.