@ncrypter90 2020-04-15
@codeofalltrades Could you let me know how can i test the in the wallet?

i have the wallet fully synced with 727 binary. Monitored the debug log with wallet unlocked, but did not find anything related to the fix.