@crypt0-fusi0n 2020-04-19
@ncrypter90 Have you observed what has been described in the testing info above from @CaveSpectre11 ?

![image](https://user-images.githubusercontent.com/48879483/79677099-d746bd80-8230-11ea-8130-77b8ae5414b4.png)