@ncrypter90 2020-04-19
@crypt0-fusi0n Yeah that's what i have tested but i dont see those entries in debug.log so i wanted to check what am i missing.