@crypt0-fusi0n 2020-04-19
Do you see **CleanBlockIndexGarbage** in the log at all? How long has the wallet been running?